moved merge sort algo to utils
